Delete others.txt
The only domains to not be blocked already outside of the YouTube shorteners, were the xcode.com website, which seems to be actually a Apple website and so, not related to Google in itself. The idea to add an others category will be redone once we actually found other misc Google Domains.