Proposed changes
Good afternoon:

First of all apologize if this is not the correct way to contribute new entries to the hosts. But this is the first time I do it this way. So I figured I would screw up, I just hope I didn't screw up to the bottom.

I apologize if this is not the right way and I have done something wrong. It wasn't my intention at all, but it's the first time I've done it this way and I probably screwed it up.

So I apologize and I beg your pardon.

A hug and thank you very much for all your time and help.