Duplicated
yt.be
youtube.de
youtube.fr
youtube.nl
youtube.pl
youtubeeducation.com
youtubegaming.com
ytimg.com

Good afternoon @nickspaargaren:

I would like to propose an idea, if I may. I don't know if a Gafam could be created for Microsoft or Apple.

And you could create a list with ads, for example from facebook, or others that are coming out, for example, in ublock origin. And they are only used to send data. And break our privacy. 

A hug and millions of thanks for your time. And I'm sorry I'm so useless in these things, it must be up to me.... 

I'm sorry, it's not my intention, but I hate Google, Apple, and others. Since privacy nothing at all